service_b.requirements.txt
Flask==3.1.2
gunicorn==23.0.0
google-cloud-logging==3.13.0
python-json-logger==4.0.0
requests==2.32.5
# open telemetry suite
opentelemetry-api>=1.29.0
opentelemetry-exporter-otlp-proto-http>=1.29.0
opentelemetry-instrumentation-flask>=0.50b0
opentelemetry-instrumentation-logging>=0.50b0
opentelemetry-instrumentation-requests>=0.50b0
opentelemetry-sdk>=1.29.0
